requirements.txt
streamlit
tensorflow-cpu==2.17.1
snscrape
certifi
numpy